requirement.txt
blinker==1.7.0
click==8.1.7
Flask==2.2.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
Werkzeug==2.2.0
pyjwt==2.0.0
datetime
uuid
Flask-SQLAlchemy==3.0.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
psycopg2-binary
SQLAlchemy-serializer
PyYAML
urllib3